backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "DRF backend with mongodb"
authors = ["jnoortheen <jnoortheen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.7"
django = "^2.1"
djangorestframework = "^3.8"
django-extensions = "^2.1"
djongo = "^1.2"

[tool.poetry.dev-dependencies]
pytest = "^3.8"
pytest-django = "^3.4"
black = "^18.3-alpha.0"
ipython = "^7.0"
pytest-cov = "^2.6"
pytest-pythonpath = "^0.7.3"
faker = "^0.9.1"
